requirements.txt
planetary-computer
pystac-client
odc-stac
stackstac
rasterio
geopandas
shapely
numpy
opencv-python
ultralytics
huggingface-hub
tqdm